% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_image.R
\name{matrix_image}
\alias{matrix_image}
\title{Draw an image plot of a matrix.}
\usage{
matrix_image(x, y, z, col = NULL, ncol = 64, ...)
}
\arguments{
\item{...}{The expected arguments \code{x}, \code{y}, and \code{z}
of the \code{graphics::image} function.}
}
\description{
This function displays an image of a matrix via
the R image function so that the image is oriented in the same
way as the matrix output. For example, if \eqn{X} is a matrix
with entries \eqn{X_{i,j}, i = 1, ..., n, j = 1, ..., m}, then
\eqn{X_{1,1}} will be displayed at the top-left, \eqn{X_{n, 1}}
the bottom-left, \eqn{X_{1, m}} the top-right, and \eqn{X_{n, m}}
the bottom-right.
}
\examples{
n <- 10
p <- 7
M <- matrix(rnorm(n * p), nrow = n)
image(M)

}
